requirements.txt
flask
opencv-python
paho-mqtt
numpy
